- JWT(JSON Web Token) implementation

+++ laravel-rest/app/Http/Controllers/MeetingController.php
@@ -6,11 +6,15 @@ use App\Meeting;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 
+use JWTAuth;
+
 class MeetingController extends Controller
 {
     public function __construct()
     {
-        // $this->>middleware('name');
+         $this->middleware('jwt.auth', ['only' =>[
+            'update', 'store', 'destroy'
+         ]]);
     }
 
     /**
@@ -52,13 +56,17 @@ class MeetingController extends Controller
             'title' => 'required',
             'description' => 'required',
             'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
         ]);
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time= $request->input('time');
-        $user_id = $request->input('user_id');
+        //$user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'time' => Carbon::createFromFormat('YmdHie', $time),
@@ -121,24 +129,17 @@ class MeetingController extends Controller
             'title' => 'required',
             'description' => 'required',
             'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
         ]);
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time= $request->input('time');
-        $user_id = $request->input('user_id');
-
-        $meeting = [
-            'title' => $title,
-            'description' => $description,
-            'time' => $time,
-            'user_id' => $user_id,
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
+        //$user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
 
@@ -181,6 +182,17 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::findOrFail($id);
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg' => 'User not found'], 404);
+        }
+
+        if(!$meeting->users()->where('users.id', $user->id)->first()){
+            return response()->json([
+                'msg' => 'user not registered for meeting, update not successful'
+            ], 401);
+        };
+
         $users = $meeting->users;
         $meeting->users()->detach();
         if(!$meeting->delete()){
